[OpenQASM]: Properly detect zero step in const ranges (#2715)
We were trying to detect if a step in a const range was zero by manually
inspecting the `step` expression before being const evaluated. This only
works if the `step` expr is the `0` literal, but it won't work if it is
a const-expression evaluating to zero. This PR changes that logic to
check the `step` value after being const-evaluated.

Fixes #2702