Initial QDK Python Package (#2707)
Creates a `qdk` python package that acts as a lightweight meta-package
for quantum-related python packages. This package wraps and re-exports
these other package to provide a consistent bundle of tooling for users.
This package is still experimental, but includes the packages: qsharp,
qsharp-widgets, qsharp-jupyterlab, azure-quantum, and qiskit.

The qdk package itself only contributes functions that let's users check
if certain optional features are available and to require those features
when writing python that depends on them. Other content found in the
root qdk module are re-exports of several utilities functions and
classes that get used across several different sub-modules.

---------

Co-authored-by: Bill Ticehurst <billti@microsoft.com>