Extract logical resource counts from a Q# program (#2717)
This adds a function `qsharp.logical_counts` to extract logical resource
counts from a Q# expression. This happens in the same way that
`qsharp.estimate` works, but it only returns the pre-layout logical
costs without performing physical resource estimation. As a result, this
call does not expect any resource-estimation specific parameters as
input.

This function is useful for just counting logical costs (e.g., get the
number of qubits and Toffoli gates, etc.) or to implement custom layout
overhead methods to be used with `estimate_custom`.