Add `--universal` flag to build.py (#2724)
This change adds a `--universal` Boolean flag to the build script that
is on by default. This allows `--no-universal` to be specified when
building, causing macOS builds to only build for the current system
architecture rather than both x86_65 and arm64. Locally this cuts
`--pip` build time in half for me.