Emit spec compliant QIR (#2590)
This updates the QIR code generation logic to comply with the Base and
Adaptive Profile specifications. In particular, it makes the following
updates:

 - The entry point function returns `i64`.
- Module level flags for computation extensions are updated to be
consistent with spec fixes for LLVM flags.
- QIR now includes global strings in the data section that are used as
tags for the output recording calls, with tag values chosen to be
consistent with the Azure Quantum service output translation.
- The `__quantum__qis__read_result__body` function is renamed to
`__quantum__rt__read_result`.
- Every program begins with a call to the `__quantum__rt__initialize`
function.