Enable implementation of `prune_error_budget` in custom estimation API (#2737)
This PR exposes the `prune_error_budget` function from the `Overhead` to
the custom Python classes that can be used with `estimate_custom`. The
function does not need to be implemented. If it is provided, it can
change the `budget` value to tune the error budget based on the
algorithm input. The signature is changed to return a `Result` in order
to catch possible Python errors and a new test cases is added to check
the behavior in presence and absence of this function.