Special case in round based factories for low target error rate (#2740)
In the Rust-based system implementation of resource estimation, the
round based factory returns a special case 1-to-1 trivial factory if the
required target out error is greater than the input T error rate from
the qubit model. This was not possible to model in the Python API. This
PR allows to add a function `trivial_distillation_unit` for that case.