Disable circuit tracing in Python by default (#2755)
The `dump_circuit` API dumps the "circuit" that was simulated in prior
calls to `qsharp.eval` or `%%qsharp` notebook cells. It's essentially a
trace of what happened so far.

The problem is, in order to make this work, we need to keep a full trace
of all the gates that have been applied, in memory, just in case someone
calls `dump_circuit`. This should be very rare. We shouldn't pay the
perf cost every time we execute Q#.

Note : `dump_circuit` is distinct from `circuit`. The latter is
untouched in this PR!


Related: #2099