Add lint warning for ambiguous if-statement followed by unary operator (#2759)
Because we support unary `+` and `-` a user can mistakenly write an
if-statement followed by something like ` - x` and expect that it
subtracts that from the result of the if-statement, when in fact parser
treats this as two separate statements. This link helps call attention
to the ambiguity and encourage the user to use clearer syntax to
disambiguate.