Automatic estimation of overhead in memory/compute architecture (#2760)
This adds a RE intrinsic called `EnableMemoryComputeArchitecture` that
emulates a compute/memory architecture with _compute qubits_ that can be
operated on with logical operations and _memory qubits_ that can be
turned into compute qubits in order to compute on them. The intrinsic
takes as parameter an upper bound `k` on the number of compute qubits.
The first `k` qubits are assumed to be compute qubits from the start.
However, when an operation is applied to a qubit that is not a compute
qubit and there is no more space for compute qubits, one needs to evict
a compute qubit and move it into memory. Similarly, if the qubit to be
acted on is in memory, it has to be moved into the compute qubit space.
The logical resource counter counts the number of times a qubit is read
from memory and written to memory. Further, one can specify an eviction
strategy (either `LeastRecentlyUsed` that evicts a compute qubit that
has been used least recently or `LeastFrequentlyUsed` that evicts a
compute qubit that has been used least frequently).