Migrate "Cancel Job" Command and Add "Delete Job" Command for testing (#2786)
This PR moves the "Cancel job" command off of the deprecated DELETE
action and onto the new POST action supported by the service. It also
adds the "Delete job" command that will make use of the DELETE action
once the service supports it. The "Delete job" is gated behind an
experimental setting until that action is fully supported by the
service. The setting is not shown in the settings UI, so it will need to
be added manually to the settings JSON to use the feature like so:
```JSON
{
    "Q#.azure.experimental.enableJobDelete": true
}
```

The Delete Job command can be used via a trash icon button on the tree
view for jobs:
<img width="367" height="172" alt="image"
src="https://github.com/user-attachments/assets/8566bfaa-9765-44a1-a71a-10d607c99c95"
/>

Or it can be used via the command palette when a job is selected in the
tree view:
<img width="464" height="92" alt="image"
src="https://github.com/user-attachments/assets/5a092470-b6c5-48eb-9237-b58590357109"
/>

And it always checks with a dialog before committing to the deletion:
<img width="319" height="110" alt="image"
src="https://github.com/user-attachments/assets/d197c553-464d-4e7d-80bb-48f9fdb7236c"
/>