Use `CCNOT` instead of `AND` for multi-controlled operations (#2807)
This avoids the measurement-based uncomputation in the intrinsic `AND`
operation in favor of just using `CCNOT` when decomposing multiple
controls on `Std.Intrinsic` gates. This leaves optimization of resulting
QIR ccx to the program author or the backend that receives the QIR. This
makes the behavior of most code more predictable and straightforward,
and avoids special (often expensive) adaptive behavior being triggered
automatically/enexpectedly by use of multi-controlled gates.
Other libraries still use the existing `AND` with it's alternate
approach, such as the arithmetic libraries, which keeps the behavior and
resource estimates of those more advanced programs largely the same.