Add ability to switch between debug and no-debug execution graph (#2809)
Currently the FIR lowerer can generate two flavors of `ExecGraph`. When
the `dbg` flag is passed to the lowerer, the `ExecGraph` contains
instructions to track block scopes. Without the `dbg` flag, these
instructions are elided for performance.

Currently the `dbg` flag is set at the initialization of the
`Interpreter`, and cannot be changed over its lifetime.

This change makes it so that we can dynamically switch between the debug
and no-debug `ExecGraph` configurations depending on what kind of
evaluation we're performing.

The motivation for this change is the following scenario:

```python
from qsharp import eval, circuit, run
eval("operation Foo() : Unit { use q = Qubit(); M(q); }")
run("Foo()")
circuit("Foo()")
```

Here, evaluating `Foo()` via the call to `circuit()` will need to use
the debug `ExecGraph` . This is because in the future, the circuit
diagrams generated with `circuit()` will show grouping of operations
based on loop block scopes. So we'll need the `ExecGraph` that contains
these block scope tracking instructions.

On the other hand, `run`, `eval` etc. need to use the no-debug
`ExecGraph` for the best performance.