Infallible `PackageId` in `hir::ItemId` (#2821)
Fixes #2815

The cause of #2815 is a recurrent issue rooted in that we don't know the
`PackageId` of packages until we insert them in the `PackageStore`. So,
we have to jump through hoops in multiple places in our codebase to
"resolve" the package ID of items, some of those resolutions are
non-trivial. We've had to fix other bugs like this in the past; this is
another instance where we didn't resolve a `PackageId` correctly.
 
All this pain is unnecessary because whenever we compile, we already
have the `PackageStore` available, so, we can ask it to give us a
`PackageId` ahead of time and use that in the compiler, so that all
items are compiled with an infallible package ID and not an
`Option::None` that we need to resolve later. That is what this PR does.

As an extra, `ItemId` is now 16 bytes instead of 24. This saves ~0.5 MB
of memory when compiling 'core + stdlib', as shown in the Memory Report.